dt is now in float and in seconds

+++ src/ntshengn_graphics_module.h
@@ -97,7 +97,7 @@ namespace NtshEngn {
 		GraphicsModule() : GraphicsModuleInterface("Neige Graphics Module") {}
 
 		void init();
-		void update(double dt);
+		void update(float dt);
 		void destroy();
 
 		// Loads the mesh described in the mesh parameter in the internal format and returns a unique identifier
